Refactor the hang-detect to work in more situations
Now we just ensure any incoming command is processed quickly. Evidence
suggests the errors I was seeing were the Google Drive API and others
timing out. The recent httpcontrol change is designed to fix that, but
this will verify.

Make the timeout 10s so httpcontrol has time to retry it's 3s timeout 3
times.